Implemented Dialogue Post-Processing and Position Serialization
The serialization to JSON now does two extra things: 1) it saves the position of the ChatNode in this program (as an (x,y) array) so we can place it properly on an import, and 2) it splits lines of dialogue into an array, as the ChatSystem expects. The export function is now functionally complete! (aside from possible minor bugs)