Nub Connection Algorithm Implemented
Algorithm for nub connection on deserialization implemented. All types of next data is gathered during deserialization, and the id of the node is mapped to a list of it's descendant ("next") node ids. Then, after all nodes are in place, we go over every node in the scene, find its descendants via this map, and connect them to each one.