BranchOnPlayerInput Missing Options Import Bug Fix
Fixed a bug where importing a player option branch node that was missing just C or both B & C would cause the missing options to still be active, which would cause an empty option on the next export if not manually fixed. The new import now checks if there is data to populate; if not, it deactivates the option in the node.