Disconnected Node Export Support + UI Tweaks
Modified the export so that nodes that are not connected to the head node are serialized after the main tree branching from the head; a warning message will be displayed to the user is this occurs. This should provide more flexibility for in-progress writing. Also modified the UI so that three update messages can appear at once, and so underscores in the export name are visible.