Import Window Implemented
Implemented the import window. It can be spawned via the old import button, and displays all possible .json files in the source directory as buttons, which can then be clicked to initialize the import. The window includes the currently loaded file name, and a warning if a file is currently loaded into the program.